feat: enhance autocommit functionality and update aliases
Updated the autocommit function to discard the temporary commit.txt file upon completion of operation. Additionally, the conclusion of the AutoCommit process is now highlighted with an echo statement. Updated the .zsh/zsh_alias to include new alias assignments. Furthermore, added validation for the 'nvim' command in .zsh/zsh_validations.